conserto no fseek
por algum motivo no meu linux tava funcionando, mas quando fui pro windows ele de um problema de memória:

glucontrol.c:137:20: warning: overflow in conversion from 'long long unsigned int' to 'long int' changes value from '18446744073709551568' to '-48' [-Woverflow]
  137 |         fseek(arq, -sizeof(struct Registro), SEEK_CUR);
      |                    ^~~~~~~~~~~~~~~~~~~~~~~~
C:/msys64/ucrt64/bin/../lib/gcc/x86_64-w64-mingw32/15.1.0/../../../../x86_64-w64-mingw32/bin/ld.exe: cannot find 1: No such file or directory
collect2.exe: error: ld returned 1 exit status

aí coloquei esse (long) que resolveu

+++ main.c
@@ -113,7 +113,7 @@ void AlterarRegistro(){
         printf("Carboidrato consumido: %d g\n", r.Carboidrato);
         printf("===============\n\n");
 
-        fseek(arq, -sizeof(struct Registro), SEEK_CUR);
+        fseek(arq, -(long)sizeof(struct Registro), SEEK_CUR);
 
         printf("----- ALTERACAO DOS DADOS -----\n");
         int DataHoraAlt[2][3] = {{0, 0, 0}, {0, 0}};
